+++ packages/contracts/contracts/LPRewards/Unipool.sol
@@ -99,7 +99,7 @@ contract Unipool is LPTokenWrapper, Ownable, CheckContract, IUnipool {
     function setParams(
         address _kumoTokenAddress,
         address _uniTokenAddress,
-        uint _duration
+        uint256 _duration
     )
         external
         override
